+++ __tests__/components/address.test.js
@@ -1,13 +1,15 @@
 import Address from '../../src/components/address'
 
 const json = {
-	line1: 'Line 1',
-	line2: 'Line 2',
-	line3: 'Line 3',
+	line_1: 'Line 1',
+	line_2: 'Line 2',
+	line_3: 'Line 3',
 	city: 'City',
 	county: 'County',
 	country: 'Country',
 	postcode: 'Postcode',
+	latitude: 52.942637,
+	longitude: -1.171216
 }
 
 const address = new Address(json);
@@ -19,3 +21,5 @@ test('city property', () => expect(address.city).toBe('City'));
 test('county property', () => expect(address.county).toBe('County'));
 test('country property', () => expect(address.country).toBe('Country'));
 test('postcode property', () => expect(address.postcode).toBe('Postcode'));
+test('latitude property', () => expect(address.latitude).toBe(52.942637));
+test('longitude property', () => expect(address.longitude).toBe(-1.171216));